requirements.txt
nicegui
requests
python-dotenv
apscheduler
pytz
line-bot-sdk